Run the broker as root
Testing shows that the broker must run as root to create a
vment_network_ref. When running a unprivileged user we fail with:

    WARN  [peer 77999] failed to create network ref: (1002) VMNET_MEM_FAILURE

This error looks like a bug and not like a forbidden operation but this
is common behavior with Apple security related calls.  In the console
we see logs about vmnet-client not running as root and not having
the "com.apple.networking" entitlement.

When running as root, the network is created successfully. This makes
sense since networks are global and it is expected that creating them
requires root or special entitlement.

Changes:
- README.md:
  - update "how it works" section
  - update "advantages compared with vmnet-helper" section
  - remove "creating user" section.
- Makefile: create the log directory as root.
- launchd plist: update to run as root.