Remove verbose serialization logs
The serialization is a dictionary but it does not reveal anything useful
about the network.

In the server we need to log the configuration, revealing the options
from the config file and the vmnet defaults.

In the client we need to start a vmnet interface with the network and
log the returned description. This will show the network configuration
and the mac address assigned by vmnet.

Example run:

    % ./vmnet-client
    INFO  requesting network
    INFO  network received

    % ./vmnet-client
    INFO  requesting network
    INFO  network received

Broker logs:

    INFO  [peer 80412] connected
    INFO  [peer 80412] creating network
    INFO  [peer 80412] send reply
    INFO  [peer 80412] disconnected
    INFO  [peer 80427] connected
    INFO  [peer 80427] send reply
    INFO  [peer 80427] disconnected