Improve client to start a vment interface
We want to test that the network we got from the broker can be used, and
show the network details. The best way to do this in a simple client is
to stat the vment interface with the network. The completion callback
includes a param xpc_dictionary_t with the network details.

The disadvantage is requiring root to run the client. In a real
application the user will start a vm using the virtualization framework
which has the permission to stat the vment interface.

Example run:

    % sudo ./vmnet-client
    INFO  connecting to broker
    INFO  requesting network
    INFO  network recived
    DEBUG starting vmnet interface from network
    INFO  vmnet interface started
    {"vmnet_mtu":1500,"vmnet_write_max_packets":256,"vmnet_read_max_packets":256,"vmnet_interface_id":"D7B57194-F3FF-49FE-A8C8-CE4B7FDDA7A5","vmnet_max_packet_size":1514,"vmnet_mac_address":"52:37:24:27:68:12"}
    INFO  waitng for termination