Use dynamic subnet
Using dynamic subnet, the created network uses the next available
subnet. After reboot, we get 192.168.2.1/24. If all clients terminate
cleanly and the broker is restarted, the broker will create the same
network again.

If the broker is stopped while clients are actively using the previous
network, the network becomes orphaned. The next time the broker is
started it will fail to create the network. If we use a dynamic subnet,
we will get the next available subnet (192.168.3.1/24), and we can
continue to work normally.

The disadvantage of dynamic subnet is not having the same IP address for
VMs after a broker is stopped when some VM are still using the network.
However with the current behaviour, orphaning a network without a way to
recover, we don't have a better way.