Do not terminate if peers are connected
Terminating the broker when peer are connected will cause the network to
become orphaned, and we will not be able to create it again. The next
time the broker is restarted, creating the same network will fail. If we
use dynamic subnet, we will get the next available subnet, which will
change the VM IP address.

To mitigate this issue, we ignore SIGTERM and SIGINT when peers are
connected. If the signal was send by launchd during shutdown, it will
send a SIGKILL after a short timeout, so we do not prevent shutdown.