broker: More robust idle shutdown
Previously we used dispatch_after since it was so simple, but adding
cancellation was ugly and created confusing logs. Replace the generation
hack with a real cancellable timer.

The implementation is much simpler:
- We stat the timer when last peer has disconnected
- We cancel the timer when the first peer is connected

Using a timer source requires lot of boilerplate since it supports so
many features that we don't need, but the details are hidden in single
function, and the rest of the code becomes simpler, and the logs are
very clear.

Example logs:

    INFO  [peer 55153] disconnected (connected peers 1)
    INFO  [peer 55165] disconnected (connected peers 0)
    DEBUG [peer 55165] ending transaction - broker can be stopped
    DEBUG [main] shutting down in 30 seconds
    INFO  [peer 55271] connected (connected peers 1)
    DEBUG [peer 55271] starting transaction to prevent termination while peers are connected
    DEBUG [peer 55271] canceling idle shutdown
    INFO  [peer 55271] send reply
    INFO  [peer 55292] connected (connected peers 2)
    INFO  [peer 55292] send reply
    INFO  [peer 55271] disconnected (connected peers 1)
    INFO  [peer 55292] disconnected (connected peers 0)
    DEBUG [peer 55292] ending transaction - broker can be stopped
    DEBUG [main] shutting down in 30 seconds
    INFO  [main] idle timeout - shutting down