lib: Add libvmentbroker client library
Add vment_broker_start_session() to vmnet-broker.h. The function
connects to the broker, ensuring that the connection remain live for the
entire lifetime of the process, and return network serialization for the
specified network name.

The test code was extracted to new test executable, using the new
library.

Example test run:

    % ./test
    INFO  received network subnet '192.168.7.1' mask '255.255.255.0' ipv6_prefix 'fd23:414a:bc10:1bb8::' prefix_len 64
    INFO  starting vmnet interface from network
    DEBUG vmnet_mtu: 1500
    DEBUG vmnet_write_max_packets: 256
    DEBUG vmnet_read_max_packets: 256
    DEBUG vmnet_interface_id: 'A0E8112C-9B24-4D2A-A0F9-4CD0C8FE7DF5'
    DEBUG vmnet_max_packet_size: 1514
    DEBUG vmnet_mac_address: '2e:b8:fc:f7:cd:61'
    INFO  vmnet interface started
    INFO  waiting for termination

NOTES:
- The broker currently ignore the network name, always returning the
  hard-coded shared network.